Moved GUI code to separate file.
